Refactor code and add docstrings to zfit.serialization.
