+++ zfit/serialization/__init__.py
@@ -1,0 +1,1 @@
+#  Copyright (c) 2020 zfit
